Workaround for gfortran v10 problems:
Scalapack (and mumps) have non-matching arguments in function calls. This has been elevated from warning to error; have added suitable flag to fortran compiler settings so revert it back to warning. This is only a temporary work-around since the autotools based build is on the way out anyway.

Also, buffered a few cases where the presence of trilinos wasn't checked.