Count nodes on outer boundary, not just a boundary
While looking at an element on the outer boundary to count the nodes
that are on the boundary, this element may have all three nodes on
boundaries since there is an internal annular boundary.
    Change this to check that a node is on the outer boundary rather
than on a boundary, as intended.

Instead of checking that the 0th and 1st node are on a boundary, check
that the 2nd node isn't on the outer boundary since the elements are
triangles.