Fix TGauss<2,16> weights and add PARANOID warning
TGauss<2,16> weights were out by a factor of four resulting in incorrect
residual contributions (only an issue when using different integration
schemes around a node).

Because the TGauss<2,16> scheme is unusual in that it has negative
weights and nodes outside the unit triangle, a PARANOID warning has been
added that warns the user of this the first time an instance of it is
made.